sub_cmd.c file added

+++ sub_cmd.c
@@ -27,7 +27,7 @@ void sub_cmd(stack_t **stack, unsigned int line_number)
 	sub_top_two =  global_variables->top_stack->next->n -
 		global_variables->top_stack->n;
 
-	global_variables->top_stack->next->n = sub_top_two;
+	global_variables->top_stack->next->n = (int) sub_top_two;
 
 	pop_cmd(stack, line_number);
 }
